+++ packages/e2e-tests/test/WalletAndNocturneContext.test.ts
@@ -56,6 +56,7 @@ describe("Wallet", async () => {
     const serverDB = open({ path: `${__dirname}/../db/localMerkleTestDB` });
     const prover = new MockSubtreeUpdateProver();
     updater = new SubtreeUpdater(wallet, serverDB, prover);
+    await updater.init();
   });
 
   async function applySubtreeUpdate() {