+++ packages/subtree-updater/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- move server to its own module
 - add tests for rapidsnark prover
 - add `SubtreeUpdater`
 - add rapidsnark prover